Struct clickhouse::Client

pub struct Client { /* private fields */ }

A client containing HTTP pool. Can be created by using Client::default() or Client::with_http_client.

Implementations

impl Client

pub fn credentials(&self) -> Credentials

returns the credentials

pub fn with_http_client(client: impl HttpClient) -> Self

Creates a new client with a specified underlying HTTP client. Now only hyper::Client is supported.

pub fn with_url(self, url: impl Into<String>) -> Self

Specifies ClickHouse’s url. Should point to HTTP endpoint.

Examples

let client = Client::default().with_url("http://localhost:8123");

pub fn with_database(self, database: impl Into<String>) -> Self

Specifies a database name.

Examples

let client = Client::default().with_database("test");

pub fn with_user(self, user: impl Into<String>) -> Self

Specifies a user.

Examples

let client = Client::default().with_user("test");

pub fn with_password(self, password: impl Into<String>) -> Self

Specifies a password.

Examples

let client = Client::default().with_password("secret");

pub fn with_compression(self, compression: Compression) -> Self

Specifies a compression mode. See Compression for details. By default, Lz4 is used.

Examples

let client = Client::default().with_compression(Compression::Lz4Hc(4));

pub fn with_option( self, name: impl Into<String>, value: impl Into<String> ) -> Self

Used to specify options that will be passed to all queries.

Example

Client::default().with_option("allow_nondeterministic_mutations", "1");

pub fn insert<T: InsertRow + Serialize>( &self, table: String ) -> Result<Insert<T>>

Starts a new INSERT statement.

Panics

If T has unnamed fields, e.g. tuples.

pub fn inserter<T: InsertRow + Serialize>( &self, table: String ) -> Result<Inserter<T>>

Creates an inserter to perform multiple INSERTs.

pub fn query(&self, query: &str) -> Query

Starts a new SELECT/DDL query.